tests complete for en passant, castling, and pawn promotion

+++ src/main/java/Data/Save.java
@@ -1,8 +1,6 @@
 package Data;
 
 import Chess.ChessGame;
-import Chess.Pieces.ChessPiece;
-import Chess.Tile;
 
 import java.io.*;
 
@@ -13,34 +11,22 @@ import static Data.FileConstants.FILE_LOCATOR;
  */
 public class Save {
 
-    public enum Tags{
-        BLACK("[B]"),
-        WHITE("[W]"),
-        BLANK("[ ]");
-        private String value;
-        Tags(String value){
-                this.value = value;
-        }
-
-
-        public String getValue() {
-            return value;
-        }
-    }
-
-
     /**
      * Clears out the current autosave to allow for a new game to write to the autosave file
      */
-// TODO: 3/22/2017 check if autosave file exists before clearing
     public static void clearAutoSave(){
         BufferedWriter writer;
+        File autoSaveFile = new File(FILE_LOCATOR.toString() + "/resources/main/AutoSave.txt");
         try {
-            writer = new BufferedWriter(new
-                    FileWriter(FILE_LOCATOR.toString() + "/resources/main/AutoSave.txt"));
-            writer.append("");
-            writer.flush();
-            writer.close();
+            if(!autoSaveFile.exists()){
+                autoSaveFile.createNewFile();
+            } else {
+                writer = new BufferedWriter(new
+                        FileWriter(FILE_LOCATOR.toString() + "/resources/main/AutoSave.txt"));
+                writer.append("");
+                writer.flush();
+                writer.close();
+            }
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -56,31 +42,9 @@ public class Save {
         BufferedWriter autoSaveFile = new BufferedWriter(
                 new FileWriter(FILE_LOCATOR.toString() + "/resources/main/AutoSave.txt",
                         true));
-        Tile[][] currentBoard = game.getBoard().getBoardArray();
-        autoSaveFile.append(game.getCurrentPlayer().name());
+
+        autoSaveFile.append(game.toString());
         autoSaveFile.newLine();
-        autoSaveFile.flush();
-        for(int i = 0; i < currentBoard.length; i++) {
-            for (int x = 0; x < currentBoard[i].length; x++) {
-                if (currentBoard[x][i] != null) {
-                    autoSaveFile.append(currentBoard[x][i].toString());
-                    if (currentBoard[x][i].getPiece().color() == ChessPiece.PieceColor.Black) {
-                        autoSaveFile.append(Tags.BLACK.getValue());
-                    } else if (currentBoard[x][i].getPiece().color() == ChessPiece.PieceColor.White) {
-                        autoSaveFile.append(Tags.WHITE.getValue());
-                    }
-                } else {
-                    autoSaveFile.append(Tags.BLANK.getValue());
-                    autoSaveFile.append(Tags.BLANK.getValue());
-
-                }
-                if (x == 7) {
-                    autoSaveFile.newLine();
-                }
-                autoSaveFile.flush();
-            }
-        }
-
         autoSaveFile.flush();
         autoSaveFile.close();
     }
